Fix code smell

+++ CoreInteractiveGraph/src/au/gov/asd/tac/constellation/graph/interaction/plugins/io/screenshot/RecentGraphScreenshotUtilities.java
@@ -162,10 +162,12 @@ public class RecentGraphScreenshotUtilities {
             filesInHistory.add(item.getFileName() + ".png");
         });
 
-        filesInDirectory.forEach(file -> {
-            if (!filesInHistory.contains(file.getName())) {
-                file.delete();
-            }
-        });
+        if (filesInDirectory != null) {
+            filesInDirectory.forEach(file -> {
+                if (!filesInHistory.contains(file.getName())) {
+                    file.delete();
+                }
+            });
+        }
     }
 }
